requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.0.0
pypdf>=4.0.0
python-docx>=1.1.0
chromadb>=0.5.0
ollama>=0.4.0
